favourite functionality added for all products and single products in fav screen

+++ lib/views/home_screen/home_screen.dart
@@ -625,7 +625,7 @@ Widget _buildProductGrid(HomeViewModel viewModel) {
         Navigator.push(context, MaterialPageRoute(builder: (_)=> DetailScreen(
           productId: productId,
           productName: productName,
-            price: int.parse(productPrice),
+           price: double.parse(productPrice).toInt(),
             discount:productDiscount,
             productImageUrls: productImageUrls, 
             description: productDescription,
